Migration update

+++ src/main/java/com/Nullmay/HoshinsLand/Entities/Account.java
@@ -6,13 +6,13 @@ import javax.persistence.*;
 @Table(name = "accounts")
 public class Account {
 
-    public Account(int id, int balance) {
+    public Account(Long id, int balance) {
         this.id = id;
         this.balance = balance;
     }
 
     @Id
-    private long id;
+    private Long id;
 
     private float balance;
 }
